Move TeacherFeedback to instructions template directory
